restructure project to support both CLI and library usage
