+++ .gitignore
@@ -1,2 +1,2 @@
-vpeak
+/vpeak
 .DS_Store
